created file for custom slider
